requirements.txt
viam-sdk
pyserial
LD2410